refactor: split cloud workspace reconciliation modules
