+++ src/gateway/worker-environments/placement-dispatch-reclaim.test.ts
@@ -1,0 +1,128 @@
+import fs from "node:fs/promises";
+import os from "node:os";
+import path from "node:path";
+import { afterEach, beforeEach, describe, expect, it } from "vitest";
+import {
+  closeOpenClawStateDatabaseForTest,
+  openOpenClawStateDatabase,
+  type OpenClawStateDatabase,
+} from "../../state/openclaw-state-db.js";
+import {
+  BUNDLE_HASH,
+  MANIFEST_REF,
+  type PlacementStore,
+  REQUEST,
+} from "./placement-dispatch-test-fixtures.js";
+import { createHarness } from "./placement-dispatch-test-harness.js";
+import { createWorkerSessionPlacementStore } from "./placement-store.js";
+
+describe("worker placement dispatch reclaim", () => {
+  let root: string;
+  let database: OpenClawStateDatabase;
+  let placementStore: PlacementStore;
+
+  beforeEach(async () => {
+    root = await fs.mkdtemp(path.join(await fs.realpath(os.tmpdir()), "openclaw-dispatch-"));
+    database = openOpenClawStateDatabase({ env: { OPENCLAW_STATE_DIR: root } });
+    placementStore = createWorkerSessionPlacementStore({ database, now: () => 1_000 });
+  });
+
+  afterEach(async () => {
+    closeOpenClawStateDatabaseForTest();
+    await fs.rm(root, { recursive: true, force: true });
+  });
+
+  it("orders the migration barrier, provisioning, sync, attachment, and activation", async () => {
+    const harness = createHarness(placementStore);
+
+    await expect(harness.service.dispatch(REQUEST)).resolves.toMatchObject({
+      state: "active",
+      environmentId: harness.ready.environmentId,
+      activeOwnerEpoch: 2,
+      workspaceBaseManifestRef: MANIFEST_REF,
+      remoteWorkspaceDir: "/worker/workspace",
+      workerBundleHash: BUNDLE_HASH,
+    });
+
+    expect(harness.log).toEqual([
+      "barrier",
+      "placement:requested",
+      "workspace",
+      "placement:provisioning",
+      "create",
+      "placement:syncing",
+      "tunnel:ready",
+      "sync",
+      "placement:starting",
+      "attach",
+      "tunnel:attached",
+      "activation",
+      "placement:active",
+    ]);
+  });
+
+  it("reconciles the workspace before destroying and reclaiming an active worker", async () => {
+    const harness = createHarness(placementStore);
+    await harness.service.dispatch(REQUEST);
+
+    await expect(
+      harness.service.reclaim({
+        sessionId: REQUEST.sessionId,
+        sessionKey: REQUEST.sessionKey,
+        agentId: REQUEST.agentId,
+      }),
+    ).resolves.toMatchObject({
+      state: "reclaimed",
+      workspaceBaseManifestRef: harness.reconciledManifestRef,
+    });
+
+    expect(harness.log.slice(-11)).toEqual([
+      "tunnel:attached",
+      "workspace:quiesce",
+      "workspace:reconcile",
+      "workspace:verify",
+      "workspace:verify-local",
+      "workspace:lease",
+      "workspace:verify",
+      "workspace:verify-local",
+      "teardown:destroy",
+      "placement:reclaimed",
+      "teardown:stop",
+    ]);
+  });
+
+  it("retains and reports cloud versions that conflict during an idle reclaim", async () => {
+    const harness = createHarness(placementStore, {
+      reconcileConflictPaths: ["src/local.ts"],
+    });
+    await harness.service.dispatch(REQUEST);
+
+    await expect(
+      harness.service.reclaim({
+        sessionId: REQUEST.sessionId,
+        sessionKey: REQUEST.sessionKey,
+        agentId: REQUEST.agentId,
+      }),
+    ).resolves.toMatchObject({ state: "reclaimed" });
+
+    expect(harness.placements.current()).toMatchObject({
+      state: "reclaimed",
+      workspaceResultConflict: {
+        paths: ["src/local.ts"],
+        stagedResultRef: expect.stringMatching(/^refs\/openclaw\/worker-results\/reclaim-/u),
+        totalCount: 1,
+      },
+    });
+
+    expect(harness.reportWorkspaceResultConflict).toHaveBeenCalledWith({
+      sessionId: REQUEST.sessionId,
+      sessionKey: REQUEST.sessionKey,
+      agentId: REQUEST.agentId,
+      paths: ["src/local.ts"],
+      stagedResultRef: expect.stringMatching(/^refs\/openclaw\/worker-results\/reclaim-/u),
+      totalCount: 1,
+    });
+    expect(placementStore.listPendingWorkspaceResults()).toEqual([]);
+    expect(harness.environments.destroy).toHaveBeenCalledOnce();
+  });
+});
